Move available layers view to more common area
