+++ mapcache/src/main/java/mil/nga/mapcache/layersprovider/LayerModel.java
@@ -1,4 +1,4 @@
-package mil.nga.mapcache.wizards.createtile;
+package mil.nga.mapcache.layersprovider;
 
 import java.util.ArrayList;
 import java.util.List;
